who loves signed bytes? I love signed bytes

+++ app/src/main/java/com/brianledbetter/kwplogger/KWP2000/MeasurementValueParser.java
@@ -4,5 +4,5 @@ package com.brianledbetter.kwplogger.KWP2000;
  * Created by b3d on 12/6/15.
  */
 public interface MeasurementValueParser {
-    public MeasurementValue parseBytes(byte[] bytes);
+    public MeasurementValue parseBytes(int[] bytes); // we use int because of wonder, wonderful signed bytes
 }
